Update AnimationBuilder and others in the same module
- Move `combinator/state` module into a file.
- Move all animation combinators into `combinator/animation_combinators` module.
- Add private `time_runner_or_default` method to reduce boilerplate.
- Add more (forgotton) methods to `AnimationBuilder`
  - `skipped`
  - `disabled` which combines `skipped` and `paused`
  - `time_scale`
  - `direction`
- Update CHANGELOG.md and also include changes from previous non PR commits.
- Update docs for some combinators
